coordinator: move FakeL2MessageService to shared package
